Fixed a bug in vortex_recenter.py in the 'for thi in range' block. The two np.where(np.logical_and statements had there dimensions switched and actually were slicing curr_weighted_dif incorrectly. If one used a non-square grid, this becomes apparent. This was causing weird values in the curr_weighted_dif array with nan's in some sectors which is not accurate.
Other part of routine where the np.where(np.logical_and... statments were being used ( I believe this was in the 'for ri in range...' block) was changed as well. All that is needed is a np.logical_and statment to slice up the array curr_vt.

The the arrays: angle_dif, weighted_dif, and obs_angle that were initialized in the beginning of the code were removed, as they were never actually used.

Lastly, the if statment after the if n>=1 statment was removed, as it said, "if n>=0" which is always true